modify:

+++ mode_selecter.h
@@ -71,6 +71,6 @@ void waypoint(Autopilot_Interface &autopilot_interface);
 void print_msg_test(Autopilot_Interface &autopilot_interface);
 void mode_circle(Autopilot_Interface &autopilot_interface);
 void do_set_mode(Autopilot_Interface &autopilot_interface,int mode_number);
-void move_ned_duration(Autopilot_Interface &autopilot_interface,float vn,float ve,float vd,int duration);
+void move_ned_duration(Autopilot_Interface &autopilot_interface,float vn,float ve,float vd,float duration);
 
 #endif // MODE_SELECTER_H_
